### 2017-12-08 #491 @whedon
[PRE REVIEW]: Vizualizing DNA sequence co-linearity with xmatchview
**Submitting author:** @warrenlr (<a href="http://orcid.org/0000-0002-9890-2293">Rene L. Warren</a>)
**Repository:** <a href="https://github.com/warrenlr/xmatchview" target ="_blank">https://github.com/warrenlr/xmatchview</a>
**Version:** v0.3
