### 2017-12-08 #491 @arfon
@warrenlr - looks like there's a minor issue with your `paper.md` file, I've fixed it up here: https://github.com/warrenlr/xmatchview/pull/1
