### 2017-12-08 #491 @warrenlr
thank you @arfon, I merged your fixed version
