### 2017-12-08 #491 @arfon
@warrenlr - there were a couple more issues with your paper and bibtex. I've got the paper to compile cleanly with these changes: https://github.com/warrenlr/xmatchview/pull/2
