### 2017-12-08 #491 @warrenlr
oh, great, just saw your message.. will merge this back
