### 2017-12-08 #491 @warrenlr
hello @arfon, I want to recompile the pdf. It seems stuck.
