### 2017-12-08 #491 @arfon
> @whedon generate pdf

@warrenlr - it looks like this is failing as the PDF is large and we can't use the GitHub API in the way I'm currently doing with the Whedon bot.

Here is the compiled PDF:

[10.21105.joss.00491.pdf](https://github.com/openjournals/joss-reviews/files/1543159/10.21105.joss.00491.pdf)
