### 2017-12-08 #491 @warrenlr
great.. looks good except for possible issue with the Coombe et al citation and reference.  I have re-committed a version with minor edits.
