### 2017-12-08 #491 @warrenlr
otherwise, it's good to go for review.
