### 2017-12-08 #491 @arfon
@warrenlr - I've compiled this again locally [10.21105.joss.00491.pdf](https://github.com/openjournals/joss-reviews/files/1543202/10.21105.joss.00491.pdf)
