### 2017-12-08 #491 @warrenlr
thank you @arfon. Still not sure why the Coomb reference is messed up.  The BibTex look ok to me
