### 2017-12-08 #491 @warrenlr
it shows:
Coombe, Rene L. AND Jackman, Lauren AND Warren. 2016. “Assembly of the Com- plete Sitka Spruce Chloroplast Genome Using 10x Genomics Gemcode Sequencing Data.” PLOS ONE 11 (9). Public Library of Science: 1–13. doi:10.1371/journal.pone.0163059.

but should be:

Coombe, Lauren AND Warren Rene L. AND Jackman, Shaun AND Yang, Chen AND Vandervalk, Benjamin P. AND Moore, Richard A. AND Pleasance, Stephen AND Coope, Robin J. AND Bohlmann, Joerg AND Holt, Robert A. AND Jones, Steven J. M. AND Birol, Inanc. 2016. “Assembly of the Complete Sitka Spruce Chloroplast Genome Using 10x Genomics Gemcode Sequencing Data.” PLOS ONE 11 (9). Public Library of Science: 1–13. doi:10.1371/journal.pone.0163059.

The text citation also only lists the first author "Coombe 2016" but should be "Coombe et al. 2016".  I have streamlined the bibtex somewhat, maybe a new build will fix it.
