### 2017-12-08 #491 @arfon
Does this look better? [10.21105.joss.00491.pdf](https://github.com/openjournals/joss-reviews/files/1543250/10.21105.joss.00491.pdf)
