### 2017-12-08 #491 @warrenlr
yes, that works @arfon. ready to proceed to next steps.
