### 2017-12-08 #491 @arfon
@warrenlr - this is the fix I made: https://github.com/warrenlr/xmatchview/pull/3
