### 2017-12-08 #491 @warrenlr
@arfon, I have made minor edits to the figures.  Could we please rebuild?  Thanks!
