### 2017-12-09 #491 @arfon
@pjotrp @mgymrek @biorelated - would one of you be willing to edit this submission for JOSS?
