### 2017-12-11 #491 @warrenlr
just noticed duplicated text. I have removed it and I'm now rebuilding.
