### 2017-12-11 #491 @warrenlr
last version before review.  I made change to fig3 to best show the utility of the tool
