### 2017-12-11 #491 @arfon
> this is a big file (fig 3 is now several Mb)

I think you've broken @whedon again :-). If the PDF takes longer to generate than 30 seconds the processing job will be killed by Heroku (where we host our servers). I've generated the PDF locally:
[10.21105.joss.00491.pdf](https://github.com/openjournals/joss-reviews/files/1548851/10.21105.joss.00491.pdf)
