### 2017-12-11 #491 @warrenlr
I am afraid that I don't have any reviewers in mind. I was looking at the list (https://github.com/openjournals/joss/blob/master/docs/reviewers.csv). Python/bioinformatics + genome experts will do (eg. ctb, Genomematt, rmflight, luizirber). Thank you fo handling the submission, @mgymrek.
Rene
