### 2017-12-12 #497 @mgymrek
Thanks @nmmsv for agreeing to review!
@warrenlr @arfon it looks like there are some issues compiling the PDF. It sounds like this issue was discussed in the pre-review thread. Was this resolved?
