### 2017-12-12 #497 @warrenlr
not to my knowledge.  I believe it has to be compiled manually on a local machine (Fig. 3 is rather large).  The latest and current build is here:

https://github.com/openjournals/joss-reviews/files/1548851/10.21105.joss.00491.pdf

thanks!
Rene
