### 2017-12-12 #497 @arfon
> @warrenlr @arfon it looks like there are some issues compiling the PDF. It sounds like this issue was discussed in the pre-review thread. Was this resolved?

It was not. The PDF is too big for our current approach so I had to manually produce it as @warrenlr said. The URL he shared above is the latest one.
