### 2017-12-15 #497 @warrenlr
thank you for agreeing to review, @nmmsv!  Were you able to access the pdf of the submitted manuscript?  Thanks, Rene
