### 2017-12-19 #497 @nmmsv
Hi @warrenlr , absolutely! I was able to access the pdf file.
I had some issues trying to install the software. I list them here:

1. It is required to use python 2.3 or 2.4. These releases are quite outdated (latest release from 2008) and some operating systems may not support them. I managed to find a compatible version for our linux system.

2. It is required to use windows font ttf files, which are not accessible to me (and I assume others that don't have a windows system around). Also, I'm not sure if it's possible to use windows fonts under GNU General Public license.

3. Another dependency is cross_match. My understanding is that this software is not open source. In order to use it (even for academic purposes), users need to complete a user agreement and send it alongside a unique ID to 3 different emails to get permission to use.
http://www.phrap.org/consed/consed.html#howToGet

4. In the dependencies sections, there are ambiguous mentions to some lines of the codes that need to be manually modified, as well as ambiguous words (i.e., "below").

Please let me know if you had any questions about my comments.
Best,
Nima
