### 2017-12-19 #497 @warrenlr
Thank you for your comments, @nmmsv.
What do you advise I do going forward?

In the mean time, I clarified the dependencies in the README.md:

    You will need to do the following before you can proceed:
      1) Download python2.3 or 2.4 from: http://www.python.org/  (The code may work with newer versions of python (not tested))
      2) Download the Python Imaging Library (PIL) from: http://www.pythonware.com/products/pil/
      3) Obtain true type fonts (ttf) and change the corresponding lines in xmatchview.py and xmatchview-conifer.py to reflect the location of the ttf. Alternatively, you could use fonts supplied by PIL instead (eg. ./pil/helvB12.pil)
      4) Download cross_match for academic use, see http://www.phrap.org and http://www.phrap.org/consed/consed.html#howToGet
      5) Make sure cross_match is in your $PATH

Here are some of my comments:
1. I haven't updated/tested the code with newer python releases.

2. Since I am not distributing the fonts with xmatchview, the code itself is still GPL, I think.
Beside, arial ttf fonts are available for linux:
eg.
https://github.com/JotJunior/PHP-Boleto-ZF2/tree/master/public/assets/fonts
Also, the code may be modified to use fonts that come with PIL (eg. ./pil/helvB12.pil)

3. Yes, xmv depends on cross_match as stated in the documentation/paper.  It is unfortunate that cross_match is not easier to obtain, but out of my control.

4. I have deleted these old requirements, they were for an older version of the code.

Let me know how to proceed from here.
Thank you!
Rene
