### 2017-12-21 #497 @mgymrek
Thanks for the clarifications.

echoing @nmmsv: Just scrolling through the source code and based on comments of @nmmsv, regarding fonts I see many of the paths are hard coded in the main .py file and need to be changed before running. It would be better to make these have reasonable defaults and allow paths to be changed via a command line argument. Also more info on how to download the required fonts, or if possible distributing those with the package, would make installation easier.
