### 2017-12-21 #497 @warrenlr
thank you @mgymrek and @nmmsv for the helpful comments and diligence.

I have streamlined the code and added the suggested option to provide a path to the font directory. I added some information to the README.md regarding font usage.

850e689959597cbadbada4e4ef96b05af0b2aeb2
b902c85b96afba2eab55e0786c7830c5e7d1c686

fyi - I have added a test folder with FASTA and cross_match output files for one of the FTL1 example from Fig.3.  Users can now test their install (does not require cross_match).

Thank you
Rene
