### 2018-01-04 #497 @nmmsv
Hello,
I did receive the license for cross_match. But I'll be trying to run the test script for now. There are some modifications that help the user in running the code, and also an error log that I've listed below.

1. Please make test script customizable, ask for inputs instead of setting manual paths (Also create usage and --help input to explain the requirements). For example, in runXMV.sh there are python paths to folders in your personal computer's home directory. Please explain (in usage or help) which paths should be used instead of those and provide a proper input mechanism.

2. Please avoid having a separate copy of xmatchview.py for the test folder. Right now there seems to be two separate source codes which creates confusion when edits are made.

3. After some path modifications and changes, I could run runXMV.sh, and there seems to be a problem when using Image library. I've attached the log and error of the program below.

[xmatchview_log.txt](https://github.com/openjournals/joss-reviews/files/1604719/xmatchview_log.txt)

Thank you.
Nima
