### 2018-01-08 #497 @warrenlr
thank you.
1. done.
2. created symbolic links
3. I tracked this problem (I am not seeing this on my system) and fixed according to instructions here: https://stackoverflow.com/questions/17451711/typeerror-when-resizing-an-image-with-pil-in-python
Let me know if this fixes it. If not, may have to change all calls to PIL modules (eg. from PIL import XX)
thanks
Rene
