### 2018-01-08 #497 @nmmsv
Hello,

I now realize that my previous log was truncated for some reason and didn't contain the actual error message. I tried again with the new version of the code, and I still get the following error:

`2629 out of 2635
2630 out of 2635
2631 out of 2635
2632 out of 2635
2633 out of 2635
2634 out of 2635
done.
Drawing repeats...
Traceback (most recent call last):
  File "xmatchview.py", line 800, in <module>
    main()
  File "xmatchview.py", line 795, in main
    drawRelationship(reference, query, match, scale, query_hit, mismatch, block_length, crossmatch_file, freq, reflength, leap, format, formatdict, protein, alpha, refexon, qryexon, qrylength, refnpos, qrynpos, refname, qryname, refzpos, qryzpos, fontpath)
  File "xmatchview.py", line 420, in drawRelationship
    bdraw = ImageDraw.Draw(back)
  File "/storage/nmmsv/xmatchview/Imaging-1.1.7/PIL/ImageDraw.py", line 294, in Draw
    return ImageDraw(im, mode)
  File "/storage/nmmsv/xmatchview/Imaging-1.1.7/PIL/ImageDraw.py", line 75, in __init__
    self.draw = Image.core.draw(self.im, blend)
TypeError: must be ImagingCore, not ImagingCore`

Apologies for confusing log.

Best,
Nima
