### 2018-01-08 #497 @warrenlr
Hi Nima,

I have preceded each PIL module import statement by "from PIL"
hopefully this fixes it.
Thanks,
Rene
