### 2018-01-11 #497 @nmmsv
Hello @warrenlr ,
Sorry I was busy yesterday. believe now things are running as they should. I've attached the resulting figure.

![xmv-ftl1_ss fa_vs_ftl1_pa fa rep_m10_b10_r1_c1](https://user-images.githubusercontent.com/15001608/34807018-2cf1dc12-f63b-11e7-8c7b-b2eee720c281.png)

I'll go ahead and read the documentation and let you know if I had questions.
Best,
Nima
