### 2018-01-12 #497 @nmmsv
Hi @warrenlr ,

I've read your paper, and I have a few comments based on the reviewer guidelines:
1. Please remove API explanation from paper, and add them to the documentation.
2. Test: Please create a means for the user to test proper installation and correct functionality. Ideally, we want to make sure the user can test every option and make sure they all work. Currently, there are example scripts in the test folder (which are another required item), but the user cannot confirm the results and correct functionality.

Thanks.
Best,
Nima
