### 2018-01-14 #497 @warrenlr
thank you @nmmsv

1. Could you confirm that the sentence you wish to remove from the paper and included in the readme is:
"Demo shell scripts that pipeline cross_match and xmatchview/xmatchview-conifer are included with the distribution and provided for guidance (runCompareTwoGenomesColinear.sh and runSpruceView.sh, respectively)" ?

2. I am not sure I understand what you require here, that is not already provided.  For this, I have created a test directory where users can test their install by comparing the figure they generate with the ones provided.  In the readme, I have a special section where I indicate:

> TEST xmatchview.py / xmatchview-conifer.py
At the unix prompt, once the package is installed, change directory to

cd ./test
Once you have downloaded pyhon and PIL and changed the paths to fonts in the xmatchview.py and xmatchview-conifer.py Execute:

./runXMV-conifer.sh FTL1_ss.fa_vs_FTL1_pa.fa.rep FTL1_ss.fa FTL1_pa.fa 200 10 1 FTL1_ss.txt FTL1_pa.txt
and
./runXMV.sh FTL1_ss.fa_vs_FTL1_pa.fa.rep FTL1_ss.fa FTL1_pa.fa 200 10 1 FTL1_ss.txt FTL1_pa.txt
**If all went well, images such as those provided in the test folder should be generated**

they can confirm results by comparing the resulting images. The functionality / parameters can be tested on these examples if need be.

Final comment:

At this point, with your insightful comments (which I am grateful for), xmatchview has been tested and demonstrated to work in your hands. I have also updated the readme since the original submission to clarify dependencies, installation and usability, thanks to your comments and that of the Editor, @mgymrek. The manuscript describes the utility of the tool, and with supporting citations, presents use cases. I trust that this is now sufficient, unless there are specifics you wish to see addressed, edited, removed, or added.  Regards, Rene
