### 2018-01-15 #497 @warrenlr
I made the requested changes.  Please see commit:

https://github.com/warrenlr/xmatchview/commit/3e1f5eb4db02a9ffc90763a252a168e15dc0ba4c
