### 2018-01-16 #497 @nmmsv
Hi @warrenlr ,

1. I was referring to the part that you explain in detail how the software works and each input feature and flag. Based on the guidelines, I think the paper should be more of an overview of motivation, high level functionality, and some applications:
> Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.
2. Thanks, now that the test procedure is explained in the documentation, it is more clear.

Best,
Nima
