### 2018-01-17 #497 @warrenlr
Hello @nmmsv ,

1. got it.  I have made the recommended changes in the latest commit.
https://github.com/warrenlr/xmatchview/commit/833529e64e655826f785136830e8734eea18da91

Thanks,
Rene
