### 2018-01-17 #497 @nmmsv
Hi @warrenlr ,
Thanks for the updates.
I currently can't see the references in the paper.md file, I think they should be in the bib file? Can you please add the final version of the paper (pdf) to the repository? I think that version should contain everything.
Best,
Nima
