### 2018-01-17 #497 @warrenlr
Trying to re-generate the pdf, hopefully won't crash the server.. @arfon , just looping you in..may need your assistance.
