### 2018-01-18 #497 @arfon
Here's the latest compiled PDF: [10.21105.joss.00497.pdf](https://github.com/openjournals/joss-reviews/files/1641309/10.21105.joss.00497.pdf)
