### 2018-01-18 #497 @warrenlr
thank you @arfon
hi @nmmsv , I added the latest compiled version of the paper (pdf) to the repository.
thank you
Rene
