### 2018-01-18 #497 @nmmsv
Hi @warrenlr ,
Perfect, I believe all the criteria has been met!
@mgymrek mentioning you to let you know and ask if any further actions needs to be taken.

Best,
Nima
